+++ Commands/ImportReports.php
@@ -11,9 +11,12 @@ namespace Piwik\Plugins\GoogleAnalyticsImporter\Commands;
 use Piwik\Config;
 use Piwik\Container\StaticContainer;
 use Piwik\Date;
+use Piwik\Option;
 use Piwik\Plugin\ConsoleCommand;
 use Piwik\Plugin\Manager;
+use Piwik\Plugins\GoogleAnalyticsImporter\CannotProcessImportException;
 use Piwik\Plugins\GoogleAnalyticsImporter\Google\Authorization;
+use Piwik\Plugins\GoogleAnalyticsImporter\Google\GoogleAnalyticsQueryService;
 use Piwik\Plugins\GoogleAnalyticsImporter\ImportConfiguration;
 use Piwik\Plugins\GoogleAnalyticsImporter\Importer;
 use Piwik\Plugins\GoogleAnalyticsImporter\ImportLock;
@@ -58,6 +61,8 @@ class ImportReports extends ConsoleCommand
             return $this->executeImpl($input, $output);
         } catch (ImportWasCancelledException $ex) {
             $output->writeln("Import was cancelled, aborting.");
+        } catch (CannotProcessImportException $ex) {
+            $output->writeln($ex->getMessage());
         }
     }
 
@@ -80,6 +85,13 @@ class ImportReports extends ConsoleCommand
 
         $idSite = $this->getIdSite($input);
         LogToSingleFileProcessor::handleLogToSingleFileInCliCommand($idSite, $output);
+
+        $canProcessNow = $this->checkIfCanProcess();
+        if($canProcessNow['canProcess'] === false){
+            $exceededMessage = 'The import will be restarted automatically at ' . $canProcessNow['nextAvailableAt'];
+            $output->writeln($exceededMessage);
+            throw new CannotProcessImportException($exceededMessage);
+        }
 
         /** @var ImportStatus $importStatus */
         $importStatus = StaticContainer::get(ImportStatus::class);
@@ -272,7 +284,7 @@ class ImportReports extends ConsoleCommand
                         break;
                     }
                 } finally {
-                    // doing it in the finally since we can get rate limited, which will result in an exception thrown
+                    // doing it in finally since we can get rate limited, which will result in an exception thrown
                     if (!$skipArchiving) {
                         $output->writeln(LogToSingleFileProcessor::$cliOutputPrefix . "Running archiving for newly imported data...");
                         $status = $importStatus->getImportStatus($idSite);
@@ -417,5 +429,19 @@ class ImportReports extends ConsoleCommand
         } catch (\Exception $ex) {
             return false;
         }
+    }
+
+    public function checkIfCanProcess()
+    {
+        $nextAvailableAt = (int) (Option::get(GoogleAnalyticsQueryService::DELAY_OPTION_NAME));
+        if (!$nextAvailableAt) {
+            return ['canProcess' => true];
+        }
+
+        if(Date::factory('now')->getTimestamp() >= $nextAvailableAt){
+            Option::delete(GoogleAnalyticsQueryService::DELAY_OPTION_NAME);
+            return ['canProcess' => true];
+        }
+        return ['canProcess' => false, 'nextAvailableAt' => Date::factory($nextAvailableAt)->toString('Y-m-d h:i a')];
     }
 }