+++ Importer.php
@@ -25,6 +25,8 @@ use Piwik\Piwik;
 use Piwik\Plugin\Manager;
 use Piwik\Plugin\ReportsProvider;
 use Piwik\Plugins\Goals\API;
+use Piwik\Plugins\GoogleAnalyticsImporter\ApiQuotaHelper;
+use Piwik\Plugins\GoogleAnalyticsImporter\Exceptions\CloudApiQuotaExceeded;
 use Piwik\Plugins\GoogleAnalyticsImporter\Google\DailyRateLimitReached;
 use Piwik\Plugins\GoogleAnalyticsImporter\Google\GoogleAnalyticsQueryService;
 use Piwik\Plugins\GoogleAnalyticsImporter\Google\GoogleCustomDimensionMapper;
@@ -46,6 +48,7 @@ use Psr\Log\LoggerInterface;
 class Importer
 {
     const IS_IMPORTED_FROM_GA_NUMERIC = 'GoogleAnalyticsImporter_isImportedFromGa';
+    const PAGE_SIZE = 100000;
 
     /**
      * @var ReportsProvider
@@ -123,9 +126,16 @@ class Importer
      */
     private $isMainImport = true;
 
+    private $apiQuotaHelper;
+
+    /**
+     * @var int
+     */
+    private $maxAvailableQueries = 0;
+
     public function __construct(ReportsProvider $reportsProvider, \Google\Service\Analytics $gaService, \Google\Service\AnalyticsReporting $gaReportingService,
                                 LoggerInterface $logger, GoogleGoalMapper $goalMapper, GoogleCustomDimensionMapper $customDimensionMapper,
-                                IdMapper $idMapper, ImportStatus $importStatus, ArchiveInvalidator $invalidator, EndDate $endDate)
+                                IdMapper $idMapper, ImportStatus $importStatus, ArchiveInvalidator $invalidator, EndDate $endDate, ApiQuotaHelper $apiQuotaHelper)
     {
         $this->reportsProvider = $reportsProvider;
         $this->gaService = $gaService;
@@ -137,6 +147,7 @@ class Importer
         $this->importStatus = $importStatus;
         $this->invalidator = $invalidator;
         $this->endDate = $endDate;
+        $this->apiQuotaHelper = $apiQuotaHelper;
     }
 
     public function setIsMainImport($isMainImport)
@@ -159,23 +170,23 @@ class Importer
 
             $startDate = Date::factory($webproperty->getCreated())->toString();
             if (!method_exists(SettingsServer::class, 'isMatomoForWordPress') || !SettingsServer::isMatomoForWordPress()) {
-                $idSite = SitesManagerAPI::getInstance()->addSite(
-                    $siteName = $webproperty->getName(),
-                    $urls = $type === \Piwik\Plugins\MobileAppMeasurable\Type::ID ? null : [$webproperty->getWebsiteUrl()],
-                    $ecommerce = $view->eCommerceTracking ? 1 : 0,
-                    $siteSearch = !empty($view->siteSearchQueryParameters),
-                    $searchKeywordParams = $view->siteSearchQueryParameters,
-                    $searchCategoryParams = $view->siteSearchCategoryParameters,
-                    $excludedIps = null,
-                    $excludedParams = $view->excludeQueryParameters,
-                    $timezone = empty($timezone) ? $view->timezone : $timezone,
-                    $currency = $view->currency,
-                    $group = null,
-                    $startDate,
-                    $excludedUserAgents = null,
-                    $keepURLFragments = null,
-                    $type
-                );
+                $siteOptions = [
+                    'siteName' => $webproperty->getName(),
+                    'urls' => [$webproperty->getWebsiteUrl()],
+                    'ecommerce' => $view->eCommerceTracking ? 1 : 0,
+                    'siteSearch' => (int) !empty($view->siteSearchQueryParameters),
+                    'searchKeywordParameters' => $view->siteSearchQueryParameters,
+                    'searchCategoryParameters' => $view->siteSearchCategoryParameters,
+                    'excludedQueryParameters' => $view->excludeQueryParameters,
+                    'timezone' => empty($timezone) ? $view->timezone : $timezone,
+                    'currency' => $view->currency,
+                    'startDate' => $startDate,
+                    'type' => $type
+                ];
+                if ($type === \Piwik\Plugins\MobileAppMeasurable\Type::ID) {
+                    unset($siteOptions['urls']);
+                }
+                $idSite = Request::processRequest('SitesManager.addSite', $siteOptions);
             } else { // matomo for wordpress
                 $site = new \WpMatomo\Site();
                 $idSite = $site->get_current_matomo_site_id();
@@ -331,8 +342,15 @@ class Importer
                 continue;
             }
 
-            $idDimension = CustomDimensionsAPI::getInstance()->configureNewCustomDimension(
-                $idSite, $extraEntry['gaDimension'], $extraEntry['dimensionScope'], $active = true);
+            try {
+                $idDimension = CustomDimensionsAPI::getInstance()->configureNewCustomDimension(
+                    $idSite, $extraEntry['gaDimension'], $extraEntry['dimensionScope'], $active = true);
+            } catch (\Exception $ex) {
+                if (strpos($ex->getMessage(), 'All Custom Dimensions for website') === 0) {
+                    $this->logger->warning("Cannot map custom dimension {$customDimension['name']}: " . $ex->getMessage());
+                    continue;
+                }
+            }
 
             $this->logger->info("Created Matomo dimension for extra dimension {gaDim} as dimension{id} with scope '{scope}'.", [
                 'gaDim' => $extraEntry['gaDimension'],
@@ -376,6 +394,7 @@ class Importer
             $this->currentLock = $lock;
             $this->noDataMessageRemoved = false;
             $this->queryCount = 0;
+            $this->maxAvailableQueries = $this->apiQuotaHelper->getBalanceApiQuota();
 
             $endPlusOne = $end->addDay(1);
 
@@ -406,8 +425,15 @@ class Importer
             $this->importStatus->finishImportIfNothingLeft($idSite);
 
             unset($recordImporters);
-        } catch (DailyRateLimitReached $ex) {
-            $this->importStatus->rateLimitReached($idSite);
+
+        } catch (DailyRateLimitReached  | CloudApiQuotaExceeded $ex) {
+            if($ex instanceof CloudApiQuotaExceeded){
+                $this->apiQuotaHelper->trackEvent('Internal Quota Exception Reached','Google_Analytics_Importer');
+                $this->importStatus->cloudRateLimitReached($idSite, $ex->getMessage());
+            } else {
+                $this->apiQuotaHelper->trackEvent('Google Quota Exception Reached','Google_Analytics_Importer');
+                $this->importStatus->rateLimitReached($idSite);
+            }
             $this->logger->info($ex->getMessage());
             return true;
         } catch (MaxEndDateReached $ex) {
@@ -506,6 +532,7 @@ class Importer
      */
     private function getRecordImporters($idSite, $viewId)
     {
+        $this->apiQuotaHelper->trackEvent('Import Attempt','Google_Analytics_Importer');
         if (empty($this->recordImporters)) {
             $recordImporters = StaticContainer::get('GoogleAnalyticsImporter.recordImporters');
 
@@ -530,7 +557,20 @@ class Importer
             $this->gaServiceReporting, $viewId, $this->getGoalMapping($idSite), $idSite, $quotaUser, StaticContainer::get(GoogleQueryObjectFactory::class), $this->logger);
         $gaQuery->setOnQueryMade(function () {
             ++$this->queryCount;
+            if($this->maxAvailableQueries != -1 && ($this->queryCount > $this->maxAvailableQueries)){
+                $this->apiQuotaHelper->saveApiUsed($this->maxAvailableQueries);
+                $this->apiQuotaHelper->trackEvent('Import Cloud Quota Exceeded','Google_Analytics_Importer');
+                $importCountForTheDay = $this->apiQuotaHelper->getImportCountForTheDay();
+                $quotaCount = $this->maxAvailableQueries;
+                //if the importer runs again after throwing CloudApiQuotaExceeded, {maxAvailableQueries} will be set as 0 and wrong count in the error message will be recorded
+                if ($quotaCount < 1 && $importCountForTheDay > 0) {
+                    $quotaCount = $importCountForTheDay;
+                }
+                throw new CloudApiQuotaExceeded($quotaCount);
+            }
         });
+        $this->apiQuotaHelper->saveApiUsed($this->queryCount);
+        $this->apiQuotaHelper->trackEvent('Import Complete','Google_Analytics_Importer');
 
         $instances = [];
         foreach ($this->recordImporters as $pluginName => $className) {